David rama

+++ headers/quicksort.hpp
@@ -1,8 +1,9 @@
+// Header para la función de Quicksort Externo, que se ejecuta en el main.cpp
 #ifndef QUICKSORT_V3_ARGS_HPP
 #define QUICKSORT_V3_ARGS_HPP
 
 #include <string>
 
-int run_quicksort(const std::string& inputFile, long N_SIZE, int a, long B_SIZE_arg, long M_SIZE_arg);
+long long run_quicksort(const std::string& inputFile, long N_SIZE, int a, long B_SIZE_arg, long M_SIZE_arg);
 
-#endif // QUICKSORT_V3_ARGS_HPP
+#endif
